renamed code.cpp to code.hpp
