+++ cpp/src/cylon/status.hpp
@@ -15,7 +15,7 @@
 #ifndef CYLON_STATUS_H
 #define CYLON_STATUS_H
 #include "string"
-#include "code.cpp"
+#include "code.hpp"
 
 namespace cylon {
 class Status {
